+++ src/Automation.js
@@ -1,0 +1,7 @@
+function autoToggle(){
+    data.autoToggled ? data.autoToggled = false : data.autoToggled = true
+}
+function automate(){
+    if (data.autoToggled)
+    buyMaxDeriv()
+}