+++ src/Legends.js
@@ -1,0 +1,3 @@
+function unlockLegends(){
+    if (data.upgrades[4].amt.gte(1)) data.hasLegend[0] = true
+}